Update create staking pool argument

+++ canisters/os/src/api/registry_staking_pool.rs
@@ -10,7 +10,7 @@ pub(crate) fn serve(
     created_at: u64,
     name: String,
     description: String,
-    annual_interest_rate: u64,
+    annual_interest_rate: u16,
     duration_in_millisecond: u64,
     bitcoin_address: String,
 ) -> Result<StakingPoolInfo, Error> {
